Add game_state_controller as a first refactoring step
