+++ tests/test_calculator.cpp
@@ -44,41 +44,73 @@ TEST(Calculator, test_set_fen_string) {
   calc_module.interpretAndSetFen(fen);
 
   auto comp_calc_module = Calculator();
-  comp_calc_module.current_game_state.board.fields[3] = 'R';
-  comp_calc_module.current_game_state.board.fields[5] = 'R';
-  comp_calc_module.current_game_state.board.fields[6] = 'K';
-  comp_calc_module.current_game_state.board.fields[12] = 'Q';
-  comp_calc_module.current_game_state.board.fields[13] = 'P';
-  comp_calc_module.current_game_state.board.fields[14] = 'P';
-  comp_calc_module.current_game_state.board.fields[15] = 'P';
-  comp_calc_module.current_game_state.board.fields[16] = 'P';
-  comp_calc_module.current_game_state.board.fields[20] = 'B';
-  comp_calc_module.current_game_state.board.fields[21] = 'N';
-  comp_calc_module.current_game_state.board.fields[25] = 'P';
-  comp_calc_module.current_game_state.board.fields[26] = 'P';
-  comp_calc_module.current_game_state.board.fields[27] = 'P';
-  comp_calc_module.current_game_state.board.fields[28] = 'N';
-  comp_calc_module.current_game_state.board.fields[41] = 'n';
-  comp_calc_module.current_game_state.board.fields[42] = 'n';
-  comp_calc_module.current_game_state.board.fields[44] = 'p';
-  comp_calc_module.current_game_state.board.fields[48] = 'p';
-  comp_calc_module.current_game_state.board.fields[49] = 'p';
-  comp_calc_module.current_game_state.board.fields[50] = 'p';
-  comp_calc_module.current_game_state.board.fields[51] = 'q';
-  comp_calc_module.current_game_state.board.fields[52] = 'b';
-  comp_calc_module.current_game_state.board.fields[53] = 'p';
-  comp_calc_module.current_game_state.board.fields[54] = 'p';
-  comp_calc_module.current_game_state.board.fields[55] = 'p';
-  comp_calc_module.current_game_state.board.fields[56] = 'r';
-  comp_calc_module.current_game_state.board.fields[59] = 'r';
-  comp_calc_module.current_game_state.board.fields[62] = 'k';
-  comp_calc_module.current_game_state.en_passant_field = 17;
-  comp_calc_module.current_game_state.half_moves_for_draw = 0;
-  comp_calc_module.current_game_state.next_turn = NextTurn::black;
-  comp_calc_module.current_game_state.next_half_move = 30;
+  comp_calc_module.game_state_controller.current_game_state.board.fields[3] =
+      'R';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[5] =
+      'R';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[6] =
+      'K';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[12] =
+      'Q';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[13] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[14] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[15] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[16] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[20] =
+      'B';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[21] =
+      'N';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[25] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[26] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[27] =
+      'P';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[28] =
+      'N';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[41] =
+      'n';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[42] =
+      'n';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[44] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[48] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[49] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[50] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[51] =
+      'q';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[52] =
+      'b';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[53] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[54] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[55] =
+      'p';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[56] =
+      'r';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[59] =
+      'r';
+  comp_calc_module.game_state_controller.current_game_state.board.fields[62] =
+      'k';
+  comp_calc_module.game_state_controller.current_game_state.en_passant_field =
+      17;
+  comp_calc_module.game_state_controller.current_game_state
+      .half_moves_for_draw = 0;
+  comp_calc_module.game_state_controller.current_game_state.next_turn =
+      NextTurn::black;
+  comp_calc_module.game_state_controller.current_game_state.next_half_move = 30;
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
 
 TEST(Calculator, test_set_game_state_2_moves) {
@@ -94,8 +126,9 @@ TEST(Calculator, test_set_game_state_2_moves) {
   calc_module.interpretAndSetFen(fen);
   calc_module.makeMovesFromFieldStrings(moves);
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
 
 TEST(Calculator, test_set_game_state_29_moves) {
@@ -115,8 +148,9 @@ TEST(Calculator, test_set_game_state_29_moves) {
   calc_module.interpretAndSetFen(fen);
   calc_module.makeMovesFromFieldStrings(moves);
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
 
 TEST(Calculator, test_en_passant_move_black) {
@@ -131,8 +165,9 @@ TEST(Calculator, test_en_passant_move_black) {
   calc_module.interpretAndSetFen(fen);
   calc_module.makeMovesFromFieldStrings(moves);
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
 
 TEST(Calculator, test_piece_transformation_white_Q) {
@@ -148,8 +183,9 @@ TEST(Calculator, test_piece_transformation_white_Q) {
   calc_module.interpretAndSetFen(fen);
   calc_module.makeMovesFromFieldStrings(moves);
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
 
 TEST(Calculator, test_piece_transformation_black_q) {
@@ -165,6 +201,7 @@ TEST(Calculator, test_piece_transformation_black_q) {
   calc_module.interpretAndSetFen(fen);
   calc_module.makeMovesFromFieldStrings(moves);
 
-  ASSERT_TRUE(compare_game_states(calc_module.current_game_state,
-                                  comp_calc_module.current_game_state));
+  ASSERT_TRUE(compare_game_states(
+      calc_module.game_state_controller.current_game_state,
+      comp_calc_module.game_state_controller.current_game_state));
 }
